+++ core/shared/src/main/scala/spire/syntax/Ops.scala
@@ -3,10 +3,8 @@ package syntax
 
 import spire.algebra._
 import spire.algebra.lattice._
-import spire.algebra.partial._
 import spire.macros.Ops
-import spire.math.{BitString, ConvertableTo, ConvertableFrom, Interval, Rational, Number}
-import spire.util.Opt
+import spire.math.{BitString, ConvertableFrom, ConvertableTo, Interval, Number, Rational}
 
 final class EqOps[A](lhs:A)(implicit ev:Eq[A]) {
   def ===[B](rhs:B)(implicit ev: B =:= A): Boolean = macro Ops.eqv[A, B]
@@ -103,23 +101,6 @@ final class SignedOps[A:Signed](lhs: A) {
   def isSignNonZero(): Boolean = macro Ops.unop[Boolean]
   def isSignNonPositive(): Boolean = macro Ops.unop[Boolean]
   def isSignNonNegative(): Boolean = macro Ops.unop[Boolean]
-}
-
-final class SemigroupoidOps[A](lhs:A)(implicit ev:Semigroupoid[A]) {
-  def |+|? (rhs: A): Opt[A] = macro Ops.binop[A, Opt[A]]
-  def |+|?? (rhs: A): Boolean = macro Ops.binop[A, Boolean]
-}
-
-final class GroupoidCommonOps[A](lhs:A)(implicit ev:Groupoid[A]) {
-  def inverse(): A = ev.inverse(lhs)
-  def isId(implicit ev1: Eq[A]): Boolean = ev.isId(lhs)(ev1)
-}
-
-final class GroupoidOps[A](lhs:A)(implicit ev:Groupoid[A]) {
-  def leftId(): A = macro Ops.unop[A]
-  def rightId(): A = macro Ops.unop[A]
-  def |-|? (rhs: A): Opt[A] = macro Ops.binop[A, Option[A]]
-  def |-|?? (rhs: A): Boolean = macro Ops.binop[A, Boolean]
 }
 
 final class SemigroupOps[A](lhs:A)(implicit ev:Semigroup[A]) {
@@ -371,41 +352,9 @@ final class VectorSpaceOps[V](x: V) {
   def :/[F](rhs:Double)(implicit ev: VectorSpace[V, F]): V = ev.divr(x, ev.scalar.fromDouble(rhs))
 }
 
-final class InnerProductSpaceOps[V](lhs: V) {
-  def dot[F](rhs: V)(implicit ev: RealInnerProductSpace[V, F]): F =
-    macro Ops.binopWithEv[V, RealInnerProductSpace[V, F], F]
-  def ⋅[F](rhs: V)(implicit ev: RealInnerProductSpace[V, F]): F =
-    macro Ops.binopWithEv[V, RealInnerProductSpace[V, F], F]
-}
-
-final class CoordinateSpaceOps[V](v: V) {
-  def _x[F](implicit ev: CoordinateSpace[V, F]): F =
-    macro Ops.unopWithEv[CoordinateSpace[V, F], F]
-
-  def _y[F](implicit ev: CoordinateSpace[V, F]): F =
-    macro Ops.unopWithEv[CoordinateSpace[V, F], F]
-
-  def _z[F](implicit ev: CoordinateSpace[V, F]): F =
-    macro Ops.unopWithEv[CoordinateSpace[V, F], F]
-
-  def coord[F](rhs: Int)(implicit ev: CoordinateSpace[V, F]): F =
-    macro Ops.binopWithEv[Int, CoordinateSpace[V, F], F]
-
-  def dimensions[F](implicit ev: CoordinateSpace[V, F]): Int =
-    macro Ops.unopWithEv[CoordinateSpace[V, F], Int]
-}
-
 final class MetricSpaceOps[V](lhs: V) {
   def distance[F](rhs: V)(implicit ev: MetricSpace[V, F]): F =
     macro Ops.binopWithEv[V, MetricSpace[V, F], F]
-}
-
-final class NormedVectorSpaceOps[V](lhs: V) {
-  def norm[F](implicit ev: RealNormedVectorSpace[V, F]): F =
-    macro Ops.unopWithEv[RealNormedVectorSpace[V, F], F]
-
-  def normalize[F](implicit ev: RealNormedVectorSpace[V, F]): V =
-    macro Ops.unopWithEv[RealNormedVectorSpace[V, F], V]
 }
 
 final class ConvertableFromOps[A](lhs:A)(implicit ev:ConvertableFrom[A]) {
@@ -436,20 +385,6 @@ final class BitStringOps[A](lhs: A)(implicit ev: BitString[A]) {
 
   def rotateLeft(rhs: Int): A = macro Ops.binop[Int, A]
   def rotateRight(rhs: Int): A = macro Ops.binop[Int, A]
-}
-
-final class LeftPartialActionOps[G](lhs: G) {
-  def ?|+|> [P](rhs: P)(implicit ev: LeftPartialAction[P, G]): Opt[P] =
-    macro Ops.binopWithEv[P, LeftPartialAction[P, G], Opt[P]]
-  def ??|+|> [P](rhs: P)(implicit ev: LeftPartialAction[P, G]): Boolean =
-    macro Ops.binopWithEv[P, LeftPartialAction[P, G], Boolean]
-}
-
-final class RightPartialActionOps[P](lhs: P) {
-  def <|+|? [G](rhs: G)(implicit ev: RightPartialAction[P, G]): Opt[P] =
-    macro Ops.binopWithEv[G, RightPartialAction[P, G], Opt[P]]
-  def <|+|?? [G](rhs: G)(implicit ev: RightPartialAction[P, G]): Boolean =
-    macro Ops.binopWithEv[G, RightPartialAction[P, G], Boolean]
 }
 
 final class LeftActionOps[G](lhs: G) {