+++ laws/src/main/scala/spire/laws/LatticeLaws.scala
@@ -3,102 +3,81 @@ package laws
 
 import spire.algebra._
 import spire.algebra.lattice._
-import spire.implicits._
 
-import org.typelevel.discipline.Laws
+trait JoinSemilatticeLaws[A] {
+  implicit def A: JoinSemilattice[A]
 
-import org.scalacheck.{Arbitrary, Prop}
-import org.scalacheck.Prop._
+  def joinAssociative(x: A, y: A, z: A): IsEq[A] =
+    A.join(A.join(x, y), z) <=> A.join(x, A.join(y, z))
 
-object LatticeLaws {
-  def apply[A : Eq : Arbitrary] = new LatticeLaws[A] {
-    def Equ = Eq[A]
-    def Arb = implicitly[Arbitrary[A]]
-  }
+  def joinCommutative(x: A, y: A): IsEq[A] =
+    A.join(x, y) <=> A.join(y, x)
+
+  def joinIdempotent(x: A): IsEq[A] =
+    A.join(x, x) <=> x
+
+  def joinPartialOrder(x: A, y: A, po: PartialOrder[A]): IsEq[Boolean] =
+    po.lteqv(x, y) <=> po.eqv(y, A.join(x, y))
 }
 
-trait LatticeLaws[A] extends Laws {
 
-  implicit def Equ: Eq[A]
-  implicit def Arb: Arbitrary[A]
+trait MeetSemilatticeLaws[A] {
+  implicit def A: MeetSemilattice[A]
 
+  def meetAssociative(x: A, y: A, z: A): IsEq[A] =
+    A.meet(A.meet(x, y), z) <=> A.meet(x, A.meet(y, z))
 
-  def joinSemilattice(implicit A: JoinSemilattice[A]) = new LatticeProperties(
-    name = "joinSemilattice",
-    parents = Nil,
-    "join.associative" → forAll((x: A, y: A, z: A) =>
-      ((x join y) join z) === (x join (y join z))
-    ),
-    "join.commutative" → forAll((x: A, y: A) =>
-      (x join y) === (y join x)
-    ),
-    "join.idempotent" → forAll((x: A) =>
-      (x join x) === x
-    )
-  )
+  def meetCommutative(x: A, y: A): IsEq[A] =
+    A.meet(x, y) <=> A.meet(y, x)
 
-  def meetSemilattice(implicit A: MeetSemilattice[A]) = new LatticeProperties(
-    name = "meetSemilattice",
-    parents = Nil,
-    "meet.associative" → forAll((x: A, y: A, z: A) =>
-      ((x meet y) meet z) === (x meet (y meet z))
-    ),
-    "meet.commutative" → forAll((x: A, y: A) =>
-      (x meet y) === (y meet x)
-    ),
-    "meet.idempotent" → forAll((x: A) =>
-      (x meet x) === x
-    )
-  )
+  def meetIdempotent(x: A): IsEq[A] =
+    A.meet(x, x) <=> x
 
-  def lattice(implicit A: Lattice[A]) = new LatticeProperties(
-    name = "lattice",
-    parents = Seq(joinSemilattice, meetSemilattice),
-    "absorption" → forAll((x: A, y: A) =>
-      ((x join (x meet y)) === x) &&
-        ((x meet (x join y)) === x)
-    )
-  )
-
-  def boundedJoinSemilattice(implicit A: BoundedJoinSemilattice[A]) = new LatticeProperties(
-    name = "boundedJoinSemilattice",
-    parents = Seq(joinSemilattice),
-    "join.identity" → forAll((x: A) =>
-      (x join A.zero) === x && (A.zero join x) === x
-    )
-  )
-
-  def boundedMeetSemilattice(implicit A: BoundedMeetSemilattice[A]) = new LatticeProperties(
-    name = "boundedMeetSemilattice",
-    parents = Seq(meetSemilattice),
-      "meet.identity" → forAll((x: A) =>
-        (x meet A.one) === x && (A.one meet x) === x
-      )
-  )
-
-  def boundedBelowLattice(implicit A: Lattice[A] with BoundedJoinSemilattice[A]) = new LatticeProperties(
-    name = "boundedBelowLattice",
-    parents = Seq(boundedJoinSemilattice, lattice)
-  )
-
-  def boundedAboveLattice(implicit A: Lattice[A] with BoundedMeetSemilattice[A]) = new LatticeProperties(
-    name = "boundedAboveLattice",
-    parents = Seq(boundedMeetSemilattice, lattice)
-  )
-
-  def boundedLattice(implicit A: BoundedLattice[A]) = new LatticeProperties(
-    name = "boundedLattice",
-    parents = Seq(boundedJoinSemilattice, boundedMeetSemilattice, lattice)
-  )
-
-  class LatticeProperties(
-    val name: String,
-    val parents: Seq[LatticeProperties],
-    val props: (String, Prop)*
-  ) extends RuleSet {
-    val bases = Seq.empty
-  }
-
+  def meetPartialOrder(x: A, y: A, po: PartialOrder[A]): IsEq[Boolean] =
+    po.lteqv(x, y) <=> po.eqv(x, A.meet(x, y))
 }
 
-// vim: expandtab:ts=2:sw=2
+trait LatticeLaws[A] extends JoinSemilatticeLaws[A] with MeetSemilatticeLaws[A] {
+  implicit def A: Lattice[A]
+
+  def absorptionJoinMeet(x: A, y: A): IsEq[A] =
+    A.join(x, A.meet(x, y)) <=> x
+
+  def absorptionMeetJoin(x: A, y: A): IsEq[A] =
+    A.meet(x, A.join(x, y)) <=> x
+}
+
+trait BoundedJoinSemilatticeLaws[A] extends JoinSemilatticeLaws[A] {
+  implicit def A: BoundedJoinSemilattice[A]
+
+  def joinIdentity(x: A): IsEq[A] =
+    A.join(x, A.zero) <=> x
+
+  def zeroPartialOrder(x: A, po: PartialOrder[A]): IsEq[Boolean] =
+    po.lteqv(A.zero, x) <=> true
+}
+
+trait BoundedMeetSemilatticeLaws[A] extends MeetSemilatticeLaws[A] {
+  implicit def A: BoundedMeetSemilattice[A]
+
+  def meetIdentity(x: A): IsEq[A] =
+    A.meet(x, A.one) <=> x
+
+  def onePartialOrder(x: A, po: PartialOrder[A]): IsEq[Boolean] =
+    po.lteqv(x, A.one) <=> true
+}
+
+trait BoundedBelowLatticeLaws[A] extends BoundedJoinSemilatticeLaws[A] with LatticeLaws[A] {
+  implicit def A: BoundedJoinSemilattice[A] with Lattice[A]
+}
+
+trait BoundedAboveLatticeLaws[A] extends BoundedMeetSemilatticeLaws[A] with LatticeLaws[A] {
+  implicit def A: BoundedMeetSemilattice[A] with Lattice[A]
+}
+
+trait BoundedLatticeLaws[A] extends BoundedBelowLatticeLaws[A] with BoundedAboveLatticeLaws[A] {
+  implicit def A: BoundedLattice[A]
+}
+
+
+// TODO: missing the distributive lattice stuff