+++ laws/src/main/scala/spire/laws/shadows/ShadowCRng.scala
@@ -1,6 +1,6 @@
 package spire.laws.shadows
 
-import spire.algebra.{CRng, CSemiring}
+import spire.algebra.CRng
 
 trait ShadowCRng[A, S] extends CRng[Shadow[A, S]]
   with ShadowCSemiring[A, S] with ShadowAdditiveAbGroup[A, S] {