+++ laws/src/main/scala/spire/laws/shadows/ShadowCSemiring.scala
@@ -4,7 +4,6 @@ import spire.algebra.CSemiring
 
 trait ShadowCSemiring[A, S] extends CSemiring[Shadow[A, S]]
   with ShadowAdditiveCMonoid[A, S] with ShadowMultiplicativeCSemigroup[A, S] {
-  import shadowing._
   implicit def A: CSemiring[A]
   implicit def S: CSemiring[S]
 }